requirements.txt
ribs[all]==0.4.0

gym==0.18.0
numpy==1.21.4
scipy==1.7.2
pandas==1.2.3
numba==0.52.0
gin-config==0.4.0
fire==0.4.0
scikit-learn==0.24.1
cma==3.0.3
logdir==0.12.0
loguru==0.5.3
matplotlib==3.4.2
seaborn==0.12.0
alive-progress==1.6.2
threadpoolctl==2.1.0
cloudpickle==1.6.0
ruamel.yaml == 0.17.7
python-slugify==5.0.2
pingouin==0.4.0
Pillow==7.2.0
PyVirtualDisplay==2.2
opencv-python==4.5.5.64

# Mario - requires interfacing with Java.
pyjnius==1.4.1

# Debugging
pyinstrument==3.3.0
pytest==6.2.2
pytest-cov==2.10.1

# Dask
dask==2.30.0
distributed==2.30.0
click==7.1.2  # Newer click causes error with dask scheduler.
bokeh==2.2.3
jupyter-server-proxy==1.5.0

# Pytorch
-f https://download.pytorch.org/whl/cu113/torch_stable.html
torch==1.10.0+cu113
torchvision==0.11.1+cu113
torchaudio==0.10.0+cu113

# Maze
gym-minigrid==1.0.1
stable-baselines3==1.4.0
scikit-image==0.19.2
